Updated the design of the reviews in the database. Updated the design of feedback to allow it to be null by default. There is a bug with a test.
Feedback can now be null so that a potentially helpful answer does not
show up as not-helpful by default.

Reviews are now stored for each user. Reviews can be created with at
POST request assuming a previous review was not created within 2 weeks,
has valid score, and the user has asked at least 1 question.

Reviews are updated with a PUT(update-review) endpoint assuming the user
is updating a review they own, the review has valid score, and the
review exists.

The User Entity now has the ability to get all questions asked by a user
and get all reviews sent by a user.

Reviews have a created and updated time attribute.

The analytics page /average-topic-ratings endpoint is removed since the
requirement was updated.